Add tests and move some auth stuff to new module
