+++ config/config.exs
@@ -1,0 +1,5 @@
+import Config
+
+if File.exists?("config/#{config_env()}.exs") do
+  import_config "#{config_env()}.exs"
+end
