+++ test/test_helper.exs
@@ -1,1 +1,5 @@
+Mimic.copy(Req)
+Mimic.copy(Req.Request)
+Mimic.copy(TwitchAPI.AuthClient)
+
 ExUnit.start()
